Initial java_fuzz_test implementation
Adds a java_fuzz_test macro for Java fuzz tests that can be invoked with
the Java fuzzer Jazzer.

Most of the existing code remains unchanged, with the following
exceptions:

* CcFuzzingBinaryInfo loses its cc prefix as it can also be used for
  other languages without changes.

A small architectural difference between Jazzer and e.g. libFuzzer fuzz
targets is that the fuzzer itself is not linked into Java fuzz targets,
but remains a separate, stand-alone binary (the "driver"). The
java_fuzz_test macro deals with this by applying the instrumentation
transition to native library dependencies directly and creating a
fuzzing_binary_uninstrumented based on a binary shim that launches the
driver.