Add full runfiles tree support
The runfiles tree of a fuzzing_binary is now packaged correctly for
OSS-Fuzz.

For a fuzz test foo, the runfiles are resolved relative to the
foo.runfiles directory in the output .tar and will be found by the Bazel
runfiles libraries.

This commit also adds a fuzz test that crashes if it can't find its
runfile.