Use correct `cc_engine_instrumentation` in `jazzer` config
While the `cc_engine_instrumentation` settings `libfuzzer` and `jazzer`
lead to the same instrumentation, this may not remain so indefinitely.
Thus the `jazzer` and `asan-jazzer` config in `.bazelrc` should use
the `jazzer` value for the setting.

Also, even when not using a sanitizer via the `jazzer` config, native
library dependencies should be instrumented for fuzzing.

This requires updating Jazzer to the latest version as previous versions
did not correctly export certain sanitizer symbols.