Pin Bazel to 4.2.2
Since OSS-Fuzz uses Bazelisk to invoke Bazel and rules_fuzzing does not
pin a Bazel version, the latest stable release of Bazel will be used.
This can (and did, twice now) unexpectedly break the build.

We should intead pin a Bazel version and consciously update it when we
add support for a new version.