Fix native library handling (#195)
* Correctly escape native library path separator

This should have been changed when Jazzer was updated past
https://github.com/CodeIntelligenceTesting/jazzer/commit/24069c388579f54ec9872e61efa44f5e6065f838
but was missed due to missing test coverage for the case of multiple
native libraries.

When Bazel 5 was released today, the use of transitions coupled with
https://github.com/bazelbuild/bazel/commit/20061f8fb7ed95924c6cdbaaaf3d06a64edf974e
caused the NativeFuzzTest example to suddenly have two native library
dependencies, which triggered the dormant bug.

* Skip non-dynamic libraries for java.library.path

transitive_native_libraries can contain LibraryToLink instances for
static libraries and object files, which do not have to be added to the
java.library.path. Prevent analysis time failures by skipping them.