Update to Bazel 7 (#241)
* Only force clang, not clang-10

clang-10 isn't available on macOS. Before Bazel 7, this never applied
to macOS since the default toolchain used Xcode, ignoring the value of
`CC` alltogether.

Also move from `--action_env` to `--repo_env`, which doesn't leak into
all actions environment, and don't set `CXX`, which isn't read by Bazel.

* Do not use `rules_cc`

The core rules still live in Bazel and `rules_cc` in particular isn't
well maintained right now. It is better to switch to the external
rulesets when the Bzlmod migration has been completed and the rules are
starlarkified, with the external rulesets being the source of truth.

* Pin Java language and runtime version to 11

Targetting Java 8 is deprecated on a JDK 21, which Bazel 7 uses for
compilation by default.

As a side effect, this allows contributors to run all tests without a
local JDK.

* Do not depend on source directories

This has always been unsound and triggers a warning with Bazel 7.

* Disable Bzlmod

Enabled by default in Bazel 7, but rules_fuzzing isn't compatible with
it yet.

* Update to Bazel 7

This requires updating some dependencies.